Reduced the amount of data transfer and sped up the processing of frames by storing the stacks in the main thread as numpy arrays instead of lists of np arrays. This means there is no more conversion and the CameraGrabber only sends one frame at a time. Removing the need for the averaging to be a variable there and removing the frame grabbing.
Restructure the flow of the pause/unpause and stop/start of the camera because closing and reopening the camera was slow and unnecessary. It turns out that the camera acquisition loop was never exiting because it was being assigned "running = false" in the same function call as the main thread was starting a new method call. Now, instead, the camera is told to stop and everything is prepared for the new frames but the camera is only started when the camera emits a "ready" signal. This is ignored when paused but starts the camera in the right mode otherwise.

Added intensity and histogram plots.