Camera Grabber change: Fixing trigger issue in which first trigger occurred but second trigger was ignored. Still not sure why but SyncReadOut trigger mode is necessary instead of Edge.
Frame handling change: Frames are converted to int32 for storage handling etc to allow for subtraction and integer maths to be performed readily. They are cast to uint16 for the cv2 draw call only.

In order to have frame rate control the LED flicker rate during difference mode, the handling of exposure time control is modified. The match/case is now handled by the GUI class which actually seems clearer anyway.